% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aritimetrica.R
\name{divisao}
\alias{divisao}
\title{Function divisao}
\usage{
divisao(a, b)
}
\arguments{
\item{a}{first number.}

\item{b}{Second number.}
}
\description{
this function is good to div.
}
\examples{
divisao(1, 4)

}
