% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aritimetrica.R
\name{multiplicar}
\alias{multiplicar}
\title{Function multiplicar}
\usage{
multiplicar(a, b)
}
\arguments{
\item{a}{first number.}

\item{b}{Second number.}
}
\description{
this function is good to mult.
}
\examples{
multiplicar(1, 4)

}
