% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes.R
\name{somar}
\alias{somar}
\title{Function to math operations}
\usage{
somar(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{its another number.}
}
\description{
This function is good to basic math operations.
}
\examples{
somar(2,3)

}
