% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aritimetrica.R
\name{subtrair}
\alias{subtrair}
\title{Function subtrair}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{first number.}

\item{b}{Second number.}
}
\description{
this function is good to sub.
}
\examples{
subtrair(1, 4)

}
