% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aritmetica.R
\name{somar}
\alias{somar}
\title{Sum Function}
\usage{
somar(a, b)
}
\arguments{
\item{a}{A any number}

\item{b}{B is another random number}
}
\description{
This function is good to sum
}
\examples{
somar(5,9)

}
