% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aritimetrica.R
\name{somar}
\alias{somar}
\title{Function somar}
\usage{
somar(a, b)
}
\arguments{
\item{a}{first number.}

\item{b}{Second number.}
}
\description{
this function is good to sum.
}
\examples{
somar(1, 4)

}
