% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aritmetrica.R
\name{somar}
\alias{somar}
\title{Function to sum}
\usage{
somar(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{It is another number.}
}
\description{
This function is good to sum.
}
\examples{
sum(2, 3)

}
