% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aritmetica.R
\name{subtrair}
\alias{subtrair}
\title{Subtract Function}
\usage{
subtrair(a, b = 2)
}
\arguments{
\item{a}{A any number}

\item{b}{B is another random number}
}
\description{
This function is good to subtract
}
\examples{
subtrair(5,9)

}
