% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aritmetrica.R
\name{subtrair}
\alias{subtrair}
\title{Function to subtract}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{It is another number again.}
}
\description{
This function is good to subtract.
}
\examples{
subtrair(2, 3)

}
